+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#12](https://github.com/LaxarJS/ax-accordion-widget/issues/12): Use laxar-mocks instead of laxar-testing now.
 - [#11](https://github.com/LaxarJS/ax-accordion-widget/issues/11): Fixed the Travis config for LaxarJS 1.0.0 compatibility